requirements.txt
numpy
pandas
matplotlib
plotly
seaborn
Flask
pandas
scikit-learn
mysql-connector-python
numpy